requirements.txt
asttokens==2.4.1
attrs==23.1.0
black==23.11.0
boto3==1.28.82
botocore==1.31.82
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
comm==0.2.0
contextlib2==21.6.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
exceptiongroup==1.1.3
executing==2.0.1
filelock==3.13.1
fonttools==4.44.0
fsspec==2023.10.0
google-pasta==0.2.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mpmath==1.3.0
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pandas==2.1.2
parso==0.8.3
pathos==0.3.1
pathspec==0.11.2
pexpect==4.8.0
Pillow==10.1.0
platformdirs==3.11.0
pox==0.3.3
ppft==1.7.6.7
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pyinstrument==3.4.2
pyinstrument-cext==0.2.4
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.12.0
ruff==0.1.5
s3transfer==0.7.0
sagemaker==2.197.0
schema==0.7.5
six==1.16.0
smdebug==1.0.34
smdebug-rulesconfig==1.0.1
stack-data==0.6.3
sympy==1.12
tblib==1.7.0
tokenize-rt==5.2.0
tomli==2.0.1
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
traitlets==5.13.0
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
wcwidth==0.2.9
zipp==3.17.0
